Autofocus the note
This was removed in the past as it looked like the cause of not having the address bar selected on CTRL + T, but it was actually an issue with non-persistent background script getting started (from Innactive to Active) and then using internal chrome.tabs.update. That problem occured only in OSX and Windows. In Chrome OS, it all worked well.

Adding autofocus back, with a better implementation this time, using selection and range.